Fix MSVC C4018 warning: signed/unsigned mismatch in nbio_write()
Cast ssize_t n to size_t before comparing with size_t len. The cast
is safe because n is guaranteed non-negative at this point (the n < 0
error path was already handled above).

+++ nonblockio.c
@@ -1524,7 +1524,7 @@ nbio_write(nbio_sock_t socket, char *buf, size_t bufSize)
       nbio_error(GET_ERRNO, TCP_ERRNO);
       return -1;
     }
-    if ( n < len )
+    if ( (size_t)n < len )
     { if ( PL_handle_signals() < 0 )
       { errno = EPLEXCEPTION;
         return -1;
